refactor: unify state and alg var in equations
No longer explicitly distinguish between algebraic variables and state variables. The symbolic equation interfaces are simplified.
