+++ Solverz/__init__.py
@@ -6,8 +6,8 @@ from Solverz.sym_algebra.functions import Sign, Abs, transpose, exp, Diag, Mat_M
 from Solverz.num_api.custom_function import minmod_flag, minmod
 from Solverz.variable.variables import Vars, TimeVars, as_Vars
 from Solverz.solvers import *
-from Solverz.code_printer.make_pyfunc import made_numerical
-from Solverz.code_printer.py_printer import render_modules
+from Solverz.code_printer.python.py_printer import made_numerical
+from Solverz.code_printer.python.module_generator import render_modules
 from Solverz.code_printer.make_module import module_printer
 from Solverz.utilities.io import save, load, save_result
 from Solverz.utilities.profile import count_time
