--- Solverz/code_printer/make_pyfunc.py
@@ -1,5 +1,0 @@
-import numpy as np
-
-from Solverz.code_printer.py_printer import made_numerical
-
-
