+++ tests/test_m3b9.py
@@ -73,12 +73,12 @@ Ixg = iVar('Ixg', [0.688836021737262, 1.57988988391346, 0.817891311823357])
 Iyg = iVar('Iyg', [-0.260077644814056, 0.192406178191528, 0.173047791590276])
 Ux = iVar('Ux',
           [1.04000110267534, 1.01157932564567, 1.02160343921907,
-          1.02502063033405, 0.993215117729926, 1.01056073782038,
-          1.02360471178264, 1.01579907336413, 1.03174403980626])
+           1.02502063033405, 0.993215117729926, 1.01056073782038,
+           1.02360471178264, 1.01579907336413, 1.03174403980626])
 Uy = iVar('Uy',
           [9.38510394478286e-07, 0.165293826097057, 0.0833635520284917,
-          -0.0396760163416718, -0.0692587531054159, -0.0651191654677445,
-          0.0665507083524658, 0.0129050646926083, 0.0354351211556429])
+           -0.0396760163416718, -0.0692587531054159, -0.0651191654677445,
+           0.0665507083524658, 0.0129050646926083, 0.0354351211556429])
 
 y0 = as_Vars([delta, omega, Ixg, Iyg, Ux, Uy])
